requirements.txt
numpy
pandas
matplotlib
librosa
setuptools
scikit-learn
pillow
streamlit
streamlit-option-menu